Fix icon names, update navbar and content_container animations
Icon names were updated to be consistent with their locations in the assets directory; all should display correctly now. The navbar expansion animation wasn't working properly so it was changed to an opacity-based animation. The content containers weren't fading in quickly enough so the speed of the animation was changed to 0.5s instead of 1.4s.