Fix css breakpoint on home page and update navbar animation
The previous breakpoint on the homepage was causing the title to wrap and push all other content down on the screen, so it has been changed to hide the element instead. The navigation animation had a minor bug that was causing it to not work properly but has now been resolved.